fix: moved diamond's facets script
